Create useListViewPanelState hook
